chore: changed editable table props to generics
