+++ src/app/(root)/(course)/course/[id]/admin/table/_table/lesson.grid.client.tsx
@@ -15,7 +15,7 @@ import {
   useReactTable,
 } from "@tanstack/react-table";
 import React from "react";
-import { getColumns } from "./lesson.cells";
+import { getColumns } from "./lesson.columns";
 
 type LessonTableProps = {
   units: {
